+++ src/na/zoned/layout_synthesizer/placer/AStarPlacer.cpp
@@ -836,8 +836,27 @@ auto AStarPlacer::placeAtomsInStorageZone(
     const auto& [nearestSLM, nearestRow, nearestCol] =
         architecture_.get().nearestStorageSite(previousSLM, previousRow,
                                                previousCol);
-    const auto discreteColumnOfNearestSite =
-        discreteTargetColumns.at(std::pair{std::cref(nearestSLM), nearestCol});
+    uint8_t discreteColumnOfNearestSite = 0;
+    const auto& it = discreteTargetColumns.find(
+        std::pair{std::cref(nearestSLM), nearestCol});
+    if (it != discreteTargetColumns.end()) {
+      discreteColumnOfNearestSite = it->second;
+    } else {
+      // distance of the actual (occupied) nearest column and a column with free
+      // sites
+      int64_t minDistance = std::numeric_limits<int64_t>::max();
+      for (const auto& [SlmColumn, discreteColumn] : discreteTargetColumns) {
+        const auto& [slm, column] = SlmColumn;
+        if (slm.get() == nearestSLM.get()) {
+          if (const auto currentDistance =
+                  std::abs(static_cast<int64_t>(column) -
+                           static_cast<int64_t>(nearestCol)) < minDistance) {
+            discreteColumnOfNearestSite = discreteColumn;
+            minDistance = currentDistance;
+          }
+        }
+      }
+    }
     minDiscreteColumnOfNearestSite =
         std::min(minDiscreteColumnOfNearestSite, discreteColumnOfNearestSite);
     maxDiscreteColumnOfNearestSite =